Remove Cut move, Use values per phase

+++ Chess/Engine/Interfaces/IBoard.cs
@@ -14,7 +14,7 @@ namespace Engine.Interfaces
         bool CanDoBlackBigCastle();
         bool CanDoWhiteBigCastle();
         bool IsOpposite(BitBoard square, Piece piece);
-        int GetValue();
+        short GetValue();
         Piece GetPiece(Square cell);
         Piece GetPiece(int cell);
         bool GetPiece(Square cell, out Piece? piece);
